# backend/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DriveScore API",
    "description": "Usage-based car insurance backend",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/me/summary": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get My Summary",
        "operationId": "get_my_summary_v1_me_summary_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DriverSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/trips": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get My Trips",
        "operationId": "get_my_trips_v1_me_trips_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TripListItem"
                  },
                  "title": "Response Get My Trips V1 Me Trips Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/trips/{trip_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get My Trip Detail",
        "operationId": "get_my_trip_detail_v1_me_trips__trip_id__get",
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trip Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insurer/overview": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Insurer Overview",
        "operationId": "get_insurer_overview_v1_insurer_overview_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsurerOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insurer/drivers": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Insurer Drivers",
        "operationId": "get_insurer_drivers_v1_insurer_drivers_get",
        "parameters": [
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "score_desc",
              "title": "Sort"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InsurerDriverItem"
                  },
                  "title": "Response Get Insurer Drivers V1 Insurer Drivers Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/insurer/drivers/{driver_id}": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Insurer Driver Detail",
        "operationId": "get_insurer_driver_detail_v1_insurer_drivers__driver_id__get",
        "parameters": [
          {
            "name": "driver_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Driver Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsurerDriverDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DriverSummaryResponse": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "premium_multiplier": {
            "type": "number",
            "title": "Premium Multiplier"
          },
          "trend": {
            "type": "string",
            "title": "Trend"
          },
          "total_trips_90d": {
            "type": "integer",
            "title": "Total Trips 90D"
          },
          "total_distance_km_90d": {
            "type": "number",
            "title": "Total Distance Km 90D"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "score",
          "confidence",
          "tier",
          "premium_multiplier",
          "trend",
          "total_trips_90d",
          "total_distance_km_90d"
        ],
        "title": "DriverSummaryResponse"
      },
      "EventResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          },
          "peak_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peak G"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          }
        },
        "type": "object",
        "required": [
          "type",
          "time",
          "peak_g",
          "lat",
          "lon"
        ],
        "title": "EventResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InsurerDriverDetailResponse": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "premium_multiplier": {
            "type": "number",
            "title": "Premium Multiplier"
          },
          "event_rates": {
            "type": "object",
            "title": "Event Rates"
          },
          "trips": {
            "items": {
              "$ref": "#/components/schemas/TripListItem"
            },
            "type": "array",
            "title": "Trips"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "score",
          "confidence",
          "tier",
          "premium_multiplier",
          "event_rates",
          "trips",
          "model_version"
        ],
        "title": "InsurerDriverDetailResponse"
      },
      "InsurerDriverItem": {
        "properties": {
          "driver_id": {
            "type": "string",
            "title": "Driver Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "premium_multiplier": {
            "type": "number",
            "title": "Premium Multiplier"
          },
          "total_trips_90d": {
            "type": "integer",
            "title": "Total Trips 90D"
          },
          "total_distance_km_90d": {
            "type": "number",
            "title": "Total Distance Km 90D"
          }
        },
        "type": "object",
        "required": [
          "driver_id",
          "score",
          "confidence",
          "tier",
          "premium_multiplier",
          "total_trips_90d",
          "total_distance_km_90d"
        ],
        "title": "InsurerDriverItem"
      },
      "InsurerOverviewResponse": {
        "properties": {
          "total_drivers": {
            "type": "integer",
            "title": "Total Drivers"
          },
          "total_trips_90d": {
            "type": "integer",
            "title": "Total Trips 90D"
          },
          "tier_distribution": {
            "type": "object",
            "title": "Tier Distribution"
          },
          "average_multiplier": {
            "type": "number",
            "title": "Average Multiplier"
          }
        },
        "type": "object",
        "required": [
          "total_drivers",
          "total_trips_90d",
          "tier_distribution",
          "average_multiplier"
        ],
        "title": "InsurerOverviewResponse"
      },
...
      "TripDetailResponse": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "duration_min": {
            "type": "number",
            "title": "Duration Min"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventResponse"
            },
            "type": "array",
            "title": "Events"
          },
          "route": {
            "items": {
              "$ref": "#/components/schemas/RoutePoint"
            },
            "type": "array",
            "title": "Route"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "started_at",
          "ended_at",
          "distance_km",
          "duration_min",
          "score",
          "confidence",
          "tier",
          "events",
          "route",
          "explanation"
        ],
        "title": "TripDetailResponse"
      },
...
      "TripListItem": {
        "properties": {
          "trip_id": {
            "type": "string",
            "title": "Trip Id"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "distance_km": {
            "type": "number",
            "title": "Distance Km"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          }
        },
        "type": "object",
        "required": [
          "trip_id",
          "started_at",
          "distance_km",
          "score",
          "tier"
        ],
        "title": "TripListItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}